Hotfix

+++ lab2/controllers/postsController.js
@@ -2,26 +2,33 @@ const mongoose = require("mongoose");
 const Posts = require("../models/postsModel");
 
 // Create a new post
-const createPost = (req, res) => {
+const createPost = async (req, res) => {
   // Get body data
   const { body } = req;
 
+  // valdiate input
+  if (!body.title || !body.description) {
+    return res
+      .status(400)
+      .json({ status: false, message: "Please enter title and description" });
+  }
+
   // Insert data into db
-  Posts.insertOne({
+  await Posts.insertOne({
     title: body.title,
     description: body.description,
   });
 
-  return app.status(201).json({
+  return res.status(201).json({
     status: true,
     message: "Post added successfully",
   });
 };
 
 // Get all posts
-const getPosts = (req, res) => {
+const getPosts = async (req, res) => {
   // Get all posts
-  const posts = Posts.find({});
+  const posts = await Posts.find({}, { title: 1, description: 1 });
 
   return res.status(200).json({
     status: true,
@@ -31,12 +38,17 @@ const getPosts = (req, res) => {
 };
 
 // Get a post by ID
-const getPostByID = (req, res) => {
+const getPostByID = async (req, res) => {
   // Post id
   const postID = req.params.id;
 
+  //Validate ID
+  if (!mongoose.isValidObjectId(postID)) {
+    return res.status(400).json({ status: false, message: "ID is not valid" });
+  }
+
   // Get post by id
-  const post = Posts.findById(postID);
+  const post = await Posts.findById(postID);
 
   return res.status(200).json({
     status: true,
@@ -46,13 +58,18 @@ const getPostByID = (req, res) => {
 };
 
 // Update a post
-const updatePostByID = (req, res) => {
+const updatePostByID = async (req, res) => {
   // Post id
   const postID = req.params.id;
 
+  //Validate ID
+  if (!mongoose.isValidObjectId(postID)) {
+    return res.status(400).json({ status: false, message: "ID is not valid" });
+  }
+
   // Update db
-  Posts.updateOne(
-    { id: postID },
+  await Posts.updateOne(
+    { _id: postID },
     {
       $set: {
         title: body.title,
@@ -68,12 +85,17 @@ const updatePostByID = (req, res) => {
 };
 
 // Delete a post
-const deletePostByID = (req, res) => {
+const deletePostByID = async (req, res) => {
   // Post id
   const postID = req.params.id;
 
+  //Validate ID
+  if (!mongoose.isValidObjectId(postID)) {
+    return res.status(400).json({ status: false, message: "ID is not valid" });
+  }
+
   // Delete from db
-  Posts.findByIdAndDelete(postID);
+  await Posts.findByIdAndDelete(postID);
 
   return res.status(202).json({
     status: true,
